+++ cmd/root.go
@@ -14,13 +14,9 @@
 package cmd
 
 import (
-	"bufio"
 	"fmt"
 	"os"
-	"sync"
 
-	"github.com/maxdobeck/sauced/logger"
-	"github.com/maxdobeck/sauced/manager"
 	"github.com/spf13/cobra"
 )
 
@@ -29,28 +25,8 @@ var rootCmd = &cobra.Command{
 	Use:   "sauced",
 	Short: "Read from a YAML file at $HOME/.config/sauced.yaml and list the changes",
 	Long:  `First test to read and watch a YAML config file.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		// Get logfile
-		logfile, err := cmd.Flags().GetString("logfile")
-		if err != nil {
-			logger.Disklog.Warn("Problem retrieving logfile flag", err)
-		}
-		logger.SetupLogfile(logfile)
-
-		manager.PruneState()
-
-		var wg sync.WaitGroup
-		// read in the sc startup commands
-		file, _ := os.Open(args[0])
-		fscanner := bufio.NewScanner(file)
-		for fscanner.Scan() {
-			if fscanner.Text() != "" {
-				wg.Add(1)
-				go manager.Start(fscanner.Text(), &wg)
-			}
-		}
-		wg.Wait()
-	},
+	// Run: func(cmd *cobra.Command, args []string) {
+	// },
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
